feat: Include an adjusted function to describe the model dynamics
Create kwf_sde_adj function to adjust the sigma of the dynamics to the size of the step, linking it to the difference that should be observed between the lower and upper node of the step. Also, make the node hold the sigma used in the step and include more information in the get_description method. Use kwf_sde_adj in the KWFTreePU class.